Remove dependabot workflow (#237)
Similar to what we previously did with the CodeQL workflow, which is to
not include it in the template. GitHub provides CodeQL and Dependabot
workflows when you enable those features on a repo's settings, and that
may be a better way to set these up since it will be catered to the repo
(i.e if it is a monorepo with JS and Python, the workflow GitHub
generates will support both).