Create schema and roles in correct database (#327)
* Connect to app database rather than postgres database when creating
schema and roles

## Context
We were getting a migration error when running the flask template
migrations with the template infrastructure. The migration error says
> sqlalchemy.exc.ProgrammingError: (psycopg2.errors.InvalidSchemaName)
no schema has been selected to create in

It turns out that the role_manager that created the schema and roles
created them in the wrong database, so when the application migrations
task connected to the correct database and tried to run migrations, the
schema it was expecting to exist didn't actually exist.

This change fixes this by connecting the role manager to the correct
database. It also changes the way the role manager accesses environment
variables to fail immediately if any environment variable is missing.