Fix check for existing published image (#334)
* Pass in correct image repository name for aws ecr describe-images call
* Rename Makefile variable PROJECT_NAME to PROJECT_ROOT
* Refactor publish-release.sh configuration parameters to not rely on
build-repository infra layer outputs
* Remove Makefile from ignore list in update-template.sh

## Context
The check for existing images wasn't working, for the
platform-test-nextjs and platform-test-flask repos, so github actions
workflows would still fail when the image was already published.

The issue was that the local IMAGE_NAME was used instead of the
image_repository_name when checking for existing images. In many
projects, including platform-test, IMAGE_NAME and image_repository_name
were the same thing, since IMAGE_NAME, which is used locally for
building docker images, is defined as
`IMAGE_NAME=$PROJECT_NAME-$APP_NAME`, while image_repository_name, which
is used to define the image repository in ECR, is defined as
`image_repository_name=${project_name}-${app_name}`.

The problem is that the variable $PROJECT_NAME is incorrectly named.
While the variable name implies that it is the same as the project_name
as defined in project-config/main.tf, in actuality it's simply the name
of the project root folder, which doesn't have to be the name of the
project as configured in project-config/main.tf. For example in
platform-test-nextjs, the project name is "platform-nextjs" (we want to
keep the project name short since it's used as a prefix for infra
resources) even though the project root is "platform-test-nextjs".

These changes fixes the publish-release.sh script to correctly use the
image_repository_name, and renames the PROJECT_NAME variable in the
Makefile to PROJECT_ROOT to avoid confusion.

In addition, we refactor the publish-release.sh script to have fewer
dependencies on infra layer terraform outputs by relying only on the
app-config module and AWS CLI commands. This allows us to delete
build-repository/outputs.tf and keep the infra root modules as
standalone modules with minimal dependencies.