Setup better scanning output (#322)
## Ticket

Resolves https://github.com/navapbc/template-infra/issues/318

## Changes

- Sets up image scanning summary
- Changes Dockle action

## Context for reviewers

This makes it easier to view any image scanning findings in the workflow
summary. To do this, I needed to edit some of the jobs to write to a set
 file, and changed Dockle to one that writes to a file so we can
output it. 

## Testing

### Old screenshot

This was using JQ filters, but we pivoted away from using that

<img width="1207" alt="Screenshot 2023-06-16 at 2 12 19 PM"
src="https://github.com/navapbc/template-infra/assets/65624843/dcfa35af-7365-4ce9-9db2-7c361657467a">