Disable alibaba secret scanning (#342)
* Disable alibaba secret scanning
* Output trivy scans to table format

## Context
Vulnerability scans were failing for platform-test-nextjs, and trivy output wasn't showing in the summary.
This change moves trivy output to logs instead of github step summary and changes format
to table to be more human readable.

It also adds a trivy secret configuration file to only focus on secrets that are
likely to be relevant to projects and leave out all the rest like
facebook, linkedin, etc that are a waste of time to scan for. Projects
can always add more if they need to. So this change adds an enable list.