Fix error when re-running account setup (#365)
## Context

Currently, when re-running `make infra-set-up-account` for an account
that is already set up, the script will error when trying to import the
bucket in to the terraform state, since it already exists in the state
file. The error shown is:

> Error: Resource already managed by Terraform
>
> Terraform is already managing a remote object for
module.backend.aws_s3_bucket.tf_state. To import to this address you
> must first remove the existing object from the state.

This change updates the script to first check if the bucket already exist,
and if so, don't create it