Explain how to fix default table permissions in app (#377)
## Changes

- Revert previous change that didn't properly fix table permissions
- Add migrations route to test app to test access to a table created by
migrator
- Update migrations file to alter default privileges for tables created
by migrator role
- Update instructions to explain the gotcha

## Context

We recently merged #373 to attempt to fix an issue with table privileges
but that didn't actually solve the problem. This PR removes the previous
change and shows an example of a proper fix that needs to be implemented
in the application migrations, updates the example app's migrations, and
added instructions in the db-setup to call this gotcha out.