Add ability to check database roles configuration (#416)
* Add "check" option to role manager lambda function
* Add check-database-roles.sh script to call role manager with "check"
option
* Add make infra-check-database-roles target
* Add cluster_id output to database module
* Add db app and migrator access to role manager lambda function
* Update set-up-database.md instructions to check database roles at the
end of setup

Other changes
* Print role manager's logs when running
create-or-update-database-roles.sh

## Context

Another developer experience thing similar to
https://github.com/navapbc/template-infra/pull/415, this change allows
someone to check that the database is properly configured without having
to spin up the service layer. This change updates the role manager
lambda function to have a "check" mode that logs in to the database as
the "migrator" and "app" users (to make sure that IAM auth works), and
attempts to perform queries that on the database to check that they have
proper privileges. Specifically, the migrator user checks the search
path to make sure that new tables are created in the right schema rather
than public, it creates a table, then the app user writes something to
that table, then the migrator user drops the table. This should improve
the first time developer setup experience, and should also help with
maintenance e.g. if someone needs to refactor the database code, they
can do so and have some confidence that their changes won't break
things.