Consolidate build-repository config (#432)
## Ticket

Part 1 of 2 for #301 

## Changes

* Add ADR for consolidating infra config into app-config module
* Add account-ids-by-name.sh utility script for mapping account names to
account ids
* Add build-repository config in app-config module
* Remove configure-app-build-repository script and update make
infra-configure-app-build-repository command
* Remove build-repository's example.tfvars and terraform.tfvars files

## Context

Revisiting earlier feedback from @kyeah that config is spread out across
multiple files is error prone and from @acouch that there's a lot of
duplicate code in the configure scripts, this PR proposes a method for
consolidating the configs all into the app-config module.

This PR just changes the build-repository module. PR #42 consolidates
the infra configuration for the app/database and app/service modules.