Add storage module for uploading documents (#510)
* Add new module infra/modules/storage that creates a bucket for
uploading documents
* Add document_upload endpoint to example app that allows a user to
upload a document to the bucket
* Split example app into separate files
* Update CI infra to delete the bucket properly

## Context

It is common for applications to need a place to store documents. The
documents can be user uploaded documents to be processed by the system,
or documents from other systems that are used for ETL pipelines. This
change adds the bucket and an example app endpoint that leverages the
bucket to upload a document.

Also, the app.py file for the example app was getting long so I split it
up into separate files.