Explicitly list allowed CI/CD actions (#537)
* Add config for list of AWS services used by project
* Create GitHub Actions policy allowing access to project's AWS services
* Remove AdministratorAccess from GitHub Actions role

## Context

Up until now, for convenience we gave the GitHub Actions CI/CD role
AdministratorAccess. This change removes the AdministratorAccess policy
and instead creates a custom policy listing all the AWS services that
the project uses. This list of services is determined by a new
configuration property in the project-config module.