Add AWS defaults to container definitions to produce clean tf plan (#603)
Unclear when this started happening, it's possible it's related to when
we upgraded either terraform or when we upgraded the AWS provider, but
in any case the terraform plan in the service layer is always showing a
diff now even after applying any changes. This change adds some AWS
default parameters to container_definitions.json to cause the plan to
show a clean diff.