Move account name config into into env-config (#605)
- Add shared_account_name property to app-config
- Add account_name to env-config
- Refactor account_names_by_environment
- Update docs

## Context for reviewers

For consistency with how we map environments to networks using the
environment_config.network_name, this change refactors how we map
environments to AWS accounts to use environment_config.account_name.
There's a special case for the build-repository which is not part of any
application environment but rather shared across all environments. For
this we add a new property in app-config called shared_account_name. We
can then delete the old account_names_by_environment map, which was a
little clunky to explain.

---------

Co-authored-by: Rocket <rocketnova@users.noreply.github.com>