Skip build step if image already published (#611)
- Skip build step if image already published
- Use commit hash of input.ref for build job concurrency

## Context for reviewers

The build-and-publish worklflow always builds the image and only checks
if the image is already published when getting ready to publish. This
change moves the published image check before the build step, allowing
the workflow to skip the build step altogether if the build has already
been published.

In addition, this change fixes the build job concurrency to use the
commit hash of the commit being built, not the commit hash of the
workflow.