Check in role manager dependencies (#493)
* Check in role manager python dependencies
* Remove local exec pip3 install step for database module

## Context

A number of people were confused by the fact that the terraform plan is
never clean due to the triggers_replace = timestamp(). There were also
concerns about the dependency on pip3. This change adds the role manager
dependency to the source code, which eliminates the need for the
local-exec provisioner step and also removes the dependency on pip3. The
cost is that there is now some vendor code in the
/modules/database/role_manager/vendor folder.