Check in role manager archive (#695)
- Remove role manager dependencies from source control
- Add role manager archive to source control
- Add make command for building role manager archive
- Add documentation on how to update role manager

## Context

The “check in role manager dependencies” change in [PR
493](https://github.com/navapbc/template-infra/pull/493) does not
function as desired. The role manager lambda still shows a diff on the
source_code_hash on a clean checkout of the repo. This change removes
the archive_file data source altogether, and checks in the zip archive
directly into the codebase. Benefits include:
- We no longer need to check in the python dependencies into source
control, which is a ton of files including binary files
- There’s nothing nondeterministic for project teams, the archive file
is exactly the zip package that gets deployed to lambda

The downside is that updating the role manager is more manual. So this
change also adds a `make infra-module-database-role-manager-archive` to
rebuild the role manager package and adds instructions for doing that.