Fix access to build repository from different AWS account (#737)
## Changes

- Configure prod environment in separate prod AWS account
- Move build repository config to separate file under app-config module
- Consolidate image_repository_name config under
build_repostiory_config.name
- Add network_name, account_name, account_id, repository_arn, and
repository_url attributes to build repository config
- Replace ecr_repository data source in modules/service with
image_repository_arn and image_repository_url passed in through
variables

## Context

The FFS project discovered a bug in multi-account project setups. The
service module determined the build repository ARN and URL using a data
source, but a data source can only fetch resources in the same account.
In order to address this, we need to eliminate the data source and
construct the repository ARN and repository URL through different means.
The way we do it is by looking at which account the build repository is
in (through app-config) and getting the id of that account, which gives
us the necessary information to do this.