Perform ETL events via Step Functions (#757)
## Ticket

Resolves https://github.com/navapbc/template-infra/issues/744

## Changes

- Changes the ETL job from invoking ECS directly, to instead invoking
Step Functions which then invokes ECS
- Makes some changes to the scheduled changes to increase their
distinction from the events jobs, and the readability of their IAM roles

## Context for reviewers

This is a similar PR to
https://github.com/navapbc/template-infra/pull/745. In-fact they use
many of the same resources. To goal of this PR is to run ETL events via
Step Functions. Step Functions create a tracking layer that's not
available when invoking ECS directly. This tracking layer allows you to
see the success and failure status of your jobs.

## Testing

https://github.com/navapbc/platform-test/pull/136