Reuse app domain for domain identity (#834)
- Use service_config.domain_name for notifications-email-domain module
- DRY up usage of domain_name and hosted_zone_id in service root module
by defining local variables

## Context

We originally thought that the domain for the SES domain identity we
create had to be the same as the root of the hosted zone domain rather
than a subdomain, but we were wrong. This change causes the domain
identity to just reuse the same domain as the app's configured domain,
which simplifies things and also removes the chance of conflict between
two apps that enable notifications.