Include prefix in service root `service_name` output
The prefix was recently moved from the env-config to the service
root[1], with the effect that the `service_config.service_name` value no
longer carries the running/real service name. Special support was added
to the `bin/update-pr-environment` script to reconstruct the actual
service name, but other things also need the prefixed service name, like
`bin/deploy-release` and `bin/run-command`, and in general feels like we
should output the actual service name.

[1] ab1fda44f30077bc1084a5cabd50d9d465e21699