Nest network and within service & database modules  (#861)
## Changes

- moves the call to `module "network"` into the service module
- moves the call to `module "network"` into the database module
- makes the associated changes to the interfaces for the service and
database modules

## Context for reviewers

Nesting the network module within the service and database modules
decreases the surface area of those modules, and makes the module inputs
more intuitive for engineers with low context on how the underlying AWS
resources work. Tat is to say, it abstracts the service and database
modules away from AWS-specific implementations.

## Testing

This PR is a no-op, see:
https://github.com/navapbc/platform-test/pull/161