Generate infra service test id from timestamp (#885)
The ultimate goal is to have a way to look for stale environments
generated by the infra service tests. The way we plan on doing that is
to encode timestamp information in the id itself, but because we want
the id to be short we are base62 encoding the id.