bin/orphaned-pr-environments: Handle no PR envs and merged PRs better (#889)
When there are no PR environments for an application, `grep 'pr-'` against the
workspace names will not match anything, in which case grep returns a code of
`1`. Since `-o pipefail` is set, `grep` returning `1` will fail the pipeline,
and since `-e` is set, the pipeline will fail the script. So catch a return of
`1` from `grep` and continue on.

Also handle merged PR statuses in addition to closed, as that will sometimes be
the case.

Also set a default for `GITHUB_STEP_SUMMARY` for when running the script
locally.