Fix service module public_endpoint output when https enabled (#893)
- Set public_endpoint to use https protocol when https is enabled

Discovered this while working on e2e testing. Preview environments use
the service module public_endpoint output to define the preview
environment URL. When HTTPS is enabled we will eventually redirect from
HTTP to HTTPS. For Rails apps we already do this at the app level. This
means that the preview environment thinks it's at the HTTP endpoint but
the server quickly redirects to HTTPS. This breaks end-to-end tests
since the tests needs a consistent base URL.