Make deploy steps for an app all use the same commit
When multiple commits hit a repo in quick succession each touching a
different app, it creates a race condition as each app CD action will
attempt to resolve the `HEAD` of `main`, not the ref that actually
triggered the workflow. So have `cd-{{app_name}}.yml` use the commit
that triggered the workflow (on `push` events) if `version` is not
provided explicitly (which will be the case for dispatch events).

That alone should fix the biggest issue, but while here ensure
`build-and-publish.yml` resolves the commit itself only once in the
`get-commit-hash` job, and then uses that everywhere else. This is so if
it is invoked manually or a deploy manually invoked non-commit ref, the
commit is fixed once the build starts at least.